DFC-417: Refactor
