+++ test/browser/pages/error.js
@@ -14,20 +14,6 @@ module.exports = class PlaywrightDevPage {
     return "Sorry, there is a problem with the service";
   }
 
-  getLocalisedSomethingWentWrongMessage(lang) {
-    if (lang.toLowerCase() === "welsh") {
-      return "Mae'n ddrwg gennym, mae problem";
-    } else {
-      return "Sorry, there is a problem with the service";
-    }
-  }
-
-  async toggleLanguage(code) {
-    await this.page.click(
-      `[data-journey-click="link - click:lang-select:${code}"]`
-    );
-  }
-
   isCurrentPage() {
     return this.page.url() === this.url;
   }
